test : 자동차 생성, 이동 기능 테스트 추가
